+++ build.gradle
@@ -30,4 +30,9 @@ dependencies {
 	testImplementation 'org.springframework.kafka:spring-kafka-test'
 	compile group: 'org.postgresql', name: 'postgresql', version: '42.2.12'
 	compile group: 'org.springframework', name: 'spring-jdbc', version: '5.2.5.RELEASE'
+	compile group: 'com.zaxxer', name: 'HikariCP', version: '3.4.2'
+}
+
+bootRun {
+	System.setProperty("spring.profiles.active", "default")
 }
