+++ src/main/java/com/cedrus/patpetillo/springkafkapingpong/config/DatabaseConfig.java
@@ -1,34 +1,18 @@
 package com.cedrus.patpetillo.springkafkapingpong.config;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.jdbc.DataSourceBuilder;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.core.env.Environment;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
+import org.springframework.context.annotation.Primary;
 
 import javax.sql.DataSource;
 
 @Configuration
-@ComponentScan("com.cedrus.patpetillo.springkafkapingpong")
-@PropertySource("classpath:database.properties")
 public class DatabaseConfig {
-    @Autowired
-    Environment environment;
-
-    private final String URL = "url";
-    private final String USER = "user";
-    private final String DRIVER = "driver";
-    private final String PASSWORD = "password";
-
     @Bean
-    DataSource dataSource() {
-        DriverManagerDataSource driverManagerDataSource = new DriverManagerDataSource();
-        driverManagerDataSource.setUrl(environment.getProperty(URL));
-        driverManagerDataSource.setUsername(environment.getProperty(USER));
-        driverManagerDataSource.setPassword(environment.getProperty(PASSWORD));
-        driverManagerDataSource.setDriverClassName(environment.getProperty(DRIVER));
-        return driverManagerDataSource;
+    @ConfigurationProperties(prefix = "datasource")
+    public DataSource dataSource() {
+        return DataSourceBuilder.create().build();
     }
 }
