+++ src/main/java/com/cedrus/patpetillo/springkafkapingpong/controller/PingPongBallController.java
@@ -1,18 +1,25 @@
 package com.cedrus.patpetillo.springkafkapingpong.controller;
 
+import com.cedrus.patpetillo.springkafkapingpong.dao.PingPongBallDAO;
 import com.cedrus.patpetillo.springkafkapingpong.model.*;
 import com.cedrus.patpetillo.springkafkapingpong.stream.PingPongBallService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Profile;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Random;
+
 @Slf4j
 @RestController
+@Profile("test")
 public class PingPongBallController {
 
     private final PingPongBallService pingPongBallService;
+    @Autowired
+    private PingPongBallDAO pingPongBallDAO;
 
     @Autowired
     public PingPongBallController(PingPongBallService pingPongBallService) {
@@ -24,7 +31,7 @@ public class PingPongBallController {
         return "Ping Pong application running...";
     }
 
-    @PostMapping(value = "/serve")
+    @PostMapping("/serve")
     public ResponseEntity<ServeBallResponse> serveBall(@RequestBody ServeBallRequest serveBallRequest) { {
             log.debug("Serve ball request: {}", serveBallRequest);
         }
@@ -55,5 +62,60 @@ public class PingPongBallController {
             ServeBallResponse serveBallResponse = new ServeBallResponse(false, String.format("Invalid ping pong ball color: %s", serveBallRequest.getColor()), HttpStatus.BAD_REQUEST);
             return new ResponseEntity<>(serveBallResponse, HttpStatus.BAD_REQUEST);
         }
+    }
+
+    @Profile("test")
+    @RequestMapping("/test")
+    public void test() {
+        log.info("\nCreate a few mock ping pong balls...");
+        for (int i = 1; i < 10; i++) {
+            PingPongTarget randomTarget = getRandomTarget();
+            Color randomColor = getRandomColor();
+            PingPongBall pingPongBall = new PingPongBall(Integer.toString(i), randomTarget, randomColor);
+            pingPongBallDAO.createPingPongBall(pingPongBall);
+        }
+
+        log.info("\nList of PingPongBalls is:");
+        for (PingPongBall p : pingPongBallDAO.getAllPingPongBalls()) {
+            log.info("{}", p);
+        }
+
+        log.info("\nGet PingPongBall with ID 7");
+        PingPongBall pingPongBallById = pingPongBallDAO.getPingPongBallById("7");
+        log.info("{}", pingPongBallById);
+
+        log.info("\nDeleting PingPongBall with ID 7");
+        pingPongBallDAO.deletePingPongBall(pingPongBallById);
+
+        log.info("\nUpdate PingPongBall with ID 1");
+        PingPongBall pingPongBall1 = pingPongBallDAO.getPingPongBallById("1");
+        pingPongBall1.setColor(Color.RED);
+        pingPongBall1.setPingPongTarget(PingPongTarget.PONG);
+        pingPongBallDAO.updatePingPongBall(pingPongBall1);
+
+        log.info("\nList of PingPongBalls is:");
+        for (PingPongBall p : pingPongBallDAO.getAllPingPongBalls()) {
+            log.info("{}", p);
+        }
+
+        log.info("\nCleaning up database:");
+        for (PingPongBall p : pingPongBallDAO.getAllPingPongBalls()) {
+            log.info("\nDeleting PingPongBall: {}", p);
+            pingPongBallDAO.deletePingPongBall((p));
+        }
+
+        log.info("\nCleanup complete...");
+    }
+
+    final Color[] colors = Color.values();
+    final PingPongTarget[] pingPongTargets = PingPongTarget.values();
+    final Random random = new Random();
+
+    private Color getRandomColor() {
+        return colors[random.nextInt(colors.length)];
+    }
+
+    private PingPongTarget getRandomTarget() {
+        return pingPongTargets[random.nextInt(pingPongTargets.length)];
     }
 }
