+++ src/main/java/com/cedrus/patpetillo/springkafkapingpong/model/PingPongBall.java
@@ -15,9 +15,4 @@ public class PingPongBall {
     public void returnBall() {
         this.pingPongTarget = pingPongTarget.equals(PingPongTarget.PING) ? PingPongTarget.PONG : PingPongTarget.PING;
     }
-
-    @Override
-    public String toString() {
-        return "Ping Pong Ball: {" + "id=" + id + ", target=" + pingPongTarget + ", color='" + color + '}';
-    }
 }
