+++ src/main/java/com/cedrus/patpetillo/springkafkapingpong/model/PingPongBallMapper.java
@@ -1,10 +1,12 @@
 package com.cedrus.patpetillo.springkafkapingpong.model;
 
 import org.springframework.jdbc.core.RowMapper;
+import org.springframework.stereotype.Component;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
+@Component
 public class PingPongBallMapper implements RowMapper<PingPongBall> {
 
     public PingPongBall mapRow(ResultSet resultSet, int i) throws SQLException {
