+++ src/main/resources/application.yaml
@@ -7,4 +7,9 @@ kafka:
   bootstrapServers: localhost:9092
   kafkaAppId: ping-pong
   autoOffsetReset: latest
+datasource:
+    driverClassName: org.postgresql.Driver
+    jdbcUrl: jdbc:postgresql://127.0.0.1:5432/postgres
+    username: postgres
+    password: postgres
 logging.level.com.cedrus: DEBUG
