--- src/main/resources/database.properties
@@ -1,4 +1,0 @@
-driver=org.postgresql.Driver
-url=jdbc:postgresql://127.0.0.1:5432/postgres
-user=postgres
-password=postgres
